Add all fix

+++ shop/app/models.py
@@ -3,6 +3,7 @@ from django.contrib.auth.models import User
 import uuid
 from django.urls import reverse
 
+
 class Employee(models.Model):
     id = models.UUIDField(primary_key=True, default=uuid.uuid4, editable=False)
     name = models.CharField(max_length=50)
@@ -21,7 +22,7 @@ class Employee(models.Model):
 class Item(models.Model):
     id = models.UUIDField(primary_key=True, default=uuid.uuid4, editable=False)
     title = models.CharField(max_length=50)
-    price = models.IntegerField()
+    price = models.DecimalField(max_digits=5, decimal_places=2)
     description = models.CharField(max_length=100)
     seller = models.ForeignKey(Employee, on_delete=models.CASCADE)
     create_date = models.DateTimeField(auto_now_add=True)
@@ -40,16 +41,14 @@ class Item(models.Model):
         return str(self.seller.get_name)
 
 
-
 class Sale(models.Model):
     id = models.UUIDField(primary_key=True, default=uuid.uuid4, editable=False)
     item = models.ForeignKey(Item, on_delete=models.CASCADE)
     quantity = models.IntegerField(default=1)
-    final_price = models.IntegerField()
+    final_price = models.DecimalField(max_digits=7, decimal_places=2)
     buyer = models.ForeignKey(User, on_delete=models.CASCADE)
     seller = models.ForeignKey(Employee, on_delete=models.CASCADE)
     create_date = models.DateTimeField(auto_now_add=True)
-
 
     class Meta:
         ordering = ['-create_date']
@@ -61,7 +60,7 @@ class Sale(models.Model):
 class NewPrice(models.Model):
     id = models.UUIDField(primary_key=True, default=uuid.uuid4, editable=False)
     item = models.ForeignKey(Item, on_delete=models.CASCADE)
-    new_price = models.IntegerField()
+    new_price = models.DecimalField(max_digits=5, decimal_places=2)
     seller = models.ForeignKey(Employee, on_delete=models.CASCADE)
     changes_date = models.DateTimeField(auto_now_add=True)
 
@@ -70,3 +69,4 @@ class NewPrice(models.Model):
 
     def __str__(self):
         return str(self.item) + ", to - $" + str(self.new_price) + ", " + str(self.changes_date)
+
